Fixed broken link to TeX FAQ concerning longtable in references section of print.xtable.Rd and xtableFtable.Rd
Surrounded caption in '+ caption +' parts of expressions where captions are constructed with as.string, thus '+ as.string(caption) +' to allow the use of glue in captions. Also 'as.string' modified as in diff provided by Jeffrey Arnold, bug report #6564. Use of glue also requested by James Curran. Added example of the use of glue in a caption to the xtable Gallery.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/xtable@131 edb9625f-4e0d-4859-8d74-9fd3b1da38cb